+++ game/gamesys/Event.cpp
@@ -331,7 +331,9 @@ idEvent *idEvent::Alloc( const idEventDef *evdef, int numargs, va_list args ) {
 idEvent::CopyArgs
 ================
 */
-void idEvent::CopyArgs( const idEventDef *evdef, int numargs, va_list args, int data[ D_EVENT_MAXARGS ] ) {
+// Techyon RB: 64 bit fixes, changed int to intptr_t
+void idEvent::CopyArgs( const idEventDef *evdef, int numargs, va_list args, intptr_t data[ D_EVENT_MAXARGS ] ) {
+// Techyon END
 	int			i;
 	const char	*format;
 	idEventArg	*arg;
@@ -460,7 +462,9 @@ idEvent::ServiceEvents
 void idEvent::ServiceEvents( void ) {
 	idEvent		*event;
 	int			num;
-	int			args[ D_EVENT_MAXARGS ];
+	// Techyon RB: 64 bit fixes, changed int to intptr_t
+	intptr_t	args[ D_EVENT_MAXARGS ];
+	// Techyon END
 	int			offset;
 	int			i;
 	int			numargs;
@@ -715,7 +719,9 @@ void idEvent::Restore( idRestoreGame *savefile ) {
 		// read the args
 		savefile->ReadInt( argsize );
 		if ( argsize != event->eventdef->GetArgSize() ) {
-			savefile->Error( "idEvent::Restore: arg size (%d) doesn't match saved arg size(%d) on event '%s'", event->eventdef->GetArgSize(), argsize, event->eventdef->GetName() );
+			// Techyon RB: fixed wrong formatting
+			savefile->Error( "idEvent::Restore: arg size (%zd) doesn't match saved arg size(%zd) on event '%s'", event->eventdef->GetArgSize(), argsize, event->eventdef->GetName() );
+			// Techyon END
 		}
 		if ( argsize ) {
 			event->data = eventDataAllocator.Alloc( argsize );
